+++ lib/ransack/adapters/active_record/context.rb
@@ -312,7 +312,7 @@ module Ransack
             alias_tracker = ::ActiveRecord::Associations::AliasTracker.create(self.klass.connection, parent.table.name, [])
             jd = Polyamorous::JoinDependency.new(
               parent.base_klass,
-              parent.base_klass.arel_table,
+              parent.table,
               Polyamorous::Join.new(name, @join_type, klass),
               alias_tracker
             )
@@ -320,12 +320,11 @@ module Ransack
           else
             jd = Polyamorous::JoinDependency.new(
               parent.base_klass,
-              parent.base_klass.arel_table,
+              parent.table,
               Polyamorous::Join.new(name, @join_type, klass),
             )
             found_association = jd.instance_variable_get(:@join_root).children.last
           end
-
 
           @associations_pot[found_association] = parent
 
@@ -337,14 +336,11 @@ module Ransack
           if ::ActiveRecord::VERSION::STRING > Constants::RAILS_5_2_0
             @join_dependency.send(:construct_tables!, jd.instance_variable_get(:@join_root))
           else
-            @join_dependency.send(
-              :construct_tables!, jd.instance_variable_get(:@join_root), found_association
-            )
+            @join_dependency.send(:construct_tables!, jd.instance_variable_get(:@join_root), found_association)
           end
 
           # Leverage the stashed association functionality in AR
           @object = @object.joins(jd)
-
           found_association
         end