+++ polyamorous/lib/polyamorous/activerecord_5.2.0_ruby_2/join_association.rb
@@ -7,8 +7,7 @@ module Polyamorous
       base.class_eval { attr_reader :join_type }
     end
 
-    def initialize(reflection, children, alias_tracker, polymorphic_class = nil,
-      join_type = Arel::Nodes::InnerJoin)
+    def initialize(reflection, children, alias_tracker, polymorphic_class = nil, join_type = Arel::Nodes::InnerJoin)
       @join_type = join_type
       if polymorphic_class && ::ActiveRecord::Base > polymorphic_class
         swapping_reflection_klass(reflection, polymorphic_class) do |reflection|
@@ -23,7 +22,7 @@ module Polyamorous
     def build_constraint(klass, table, key, foreign_table, foreign_key)
       if reflection.polymorphic?
         super(klass, table, key, foreign_table, foreign_key)
-        .and(foreign_table[reflection.foreign_type].eq(reflection.klass.name))
+          .and(foreign_table[reflection.foreign_type].eq(reflection.klass.name))
       else
         super(klass, table, key, foreign_table, foreign_key)
       end