+++ polyamorous/lib/polyamorous/activerecord_5.2.1_ruby_2/join_dependency.rb
@@ -3,7 +3,6 @@
 module Polyamorous
   module JoinDependencyExtensions
     # Replaces ActiveRecord::Associations::JoinDependency#build
-    #
     def build(associations, base_klass)
       associations.map do |name, right|
         if name.is_a? Join
@@ -29,6 +28,31 @@ module Polyamorous
         end
       end
     end
+
+    def join_constraints(joins_to_add, join_type, alias_tracker)
+      @alias_tracker = alias_tracker
+
+      construct_tables!(join_root)
+      joins = make_join_constraints(join_root, join_type)
+
+      joins.concat joins_to_add.flat_map { |oj|
+        construct_tables!(oj.join_root)
+        if join_root.match?(oj.join_root) && join_root.table.name == oj.join_root.table.name
+          walk join_root, oj.join_root
+        else
+          make_join_constraints(oj.join_root, join_type)
+        end
+      }
+    end
+
+    private
+      def make_constraints(parent, child, join_type = Arel::Nodes::OuterJoin)
+        foreign_table = parent.table
+        foreign_klass = parent.base_klass
+        join_type = child.join_type || join_type if join_type == Arel::Nodes::InnerJoin
+        joins = child.join_constraints(foreign_table, foreign_klass, join_type, alias_tracker)
+        joins.concat child.children.flat_map { |c| make_constraints(child, c, join_type) }
+      end
 
     module ClassMethods
       # Prepended before ActiveRecord::Associations::JoinDependency#walk_tree