+++ polyamorous/lib/polyamorous/activerecord_5.2.1_ruby_2/reflection.rb
@@ -1,0 +1,2 @@
+# active_record_5.2.1_ruby_2/reflection.rb
+require 'polyamorous/activerecord_5.2.0_ruby_2/reflection'