+++ polyamorous/spec/blueprints/notes.rb
@@ -1,3 +1,0 @@
-Note.blueprint do
-  note
-end